requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==1.6.6
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
enum34==1.1.6
Flask==1.1.1
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
gevent==1.4.0
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.26.0
h5py==2.10.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
numpy==1.16.5
oauthlib==3.1.0
opt-einsum==2.3.2
Pillow==6.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pylint==1.9.5
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.2.2
singledispatch==3.4.0.3
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
